+++ src/Component/AbstractComponent.php
@@ -120,28 +120,11 @@ abstract class AbstractComponent
 			}
 		}
 
-		$previewData = null;
-
-		if (\is_string($data["_editable"] ?? null) && \preg_match('~^<!--#storyblok#(.*)-->$~', $data["_editable"], $matches))
-		{
-			try
-			{
-				$previewData = \json_decode(\stripslashes($matches[1]), true, flags: \JSON_THROW_ON_ERROR);
-			}
-			catch (\JsonException $exception)
-			{
-				throw new InvalidDataException(\sprintf(
-					"Encountered invalid preview data: '%s'",
-					$data["_editable"],
-				), previous: $exception);
-			}
-		}
-
 		$componentData = new ComponentData(
 			$data["_uid"],
 			static::getKey(),
 			$transformedData,
-			previewData: $previewData,
+			previewData: $data["_editable"] ?? null,
 		);
 
 		if ($dataVisitor instanceof ComponentDataVisitorInterface)