+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+3.3.0
+=====
+
+* (feature) Extract parsing of `_editable` data into `PreviewDataParser`.
+* (internal) Expose `_editable` data via `StoryMetaData::getPreviewData()`.
+* (feature) Add `ComponentPreviewData` helper to easily render component meta and preview data.
+
+
 3.2.2
 =====